pyproject.toml
[tool.poetry]
name = "hierarchial-data"
version = "0.1.0"
description = ""
authors = ["Justin_Phillips <dcoltwinchester1967@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-mptt = "^0.11.0"
django = "^3.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
